put MapPointTool in its own file/class
