+++ source/CoordinateConversion/ArcMapAddinCoordinateConversion/CoordinateConversionButton.cs
@@ -49,127 +49,6 @@ namespace ArcMapAddinCoordinateConversion
 
     public class PointTool : ESRI.ArcGIS.Desktop.AddIns.Tool
     {
-        ISnappingEnvironment m_SnappingEnv;
-        IPointSnapper m_Snapper;
-        ISnappingFeedback m_SnappingFeedback;
-
-        public PointTool()
-        {
-        }
-
-        protected override void OnUpdate()
-        {
-            Enabled = ArcMap.Application != null;
-        }
-
-        protected override void OnActivate()
-        {
-            //Get the snap environment and initialize the feedback
-            UID snapUID = new UID();
-
-            snapUID.Value = "{E07B4C52-C894-4558-B8D4-D4050018D1DA}";
-            m_SnappingEnv = ArcMap.Application.FindExtensionByCLSID(snapUID) as ISnappingEnvironment;
-            m_Snapper = m_SnappingEnv.PointSnapper;
-            m_SnappingFeedback = new SnappingFeedbackClass();
-            m_SnappingFeedback.Initialize(ArcMap.Application, m_SnappingEnv, true);
-        }
-
-        protected override void OnMouseDown(ESRI.ArcGIS.Desktop.AddIns.Tool.MouseEventArgs arg)
-        {
-            if (arg.Button != System.Windows.Forms.MouseButtons.Left)
-                return;
-            try
-            {
-                var point = GetMapPoint(arg.X, arg.Y);
-                ISnappingResult snapResult = null;
-                //Try to snap the current position
-                snapResult = m_Snapper.Snap(point);
-                m_SnappingFeedback.Update(null, 0);
-                if (snapResult != null && snapResult.Location != null)
-                    point = snapResult.Location;
-
-                //TODO replace with Mediator calls, new map point
-                //var doc = AddIn.FromID<ArcMapAddinCoordinateConversion.DockableWindowCoordinateConversion.AddinImpl>(ThisAddIn.IDs.DockableWindowCoordinateConversion);
-
-                //if (doc != null && point != null)
-                //{
-                //    doc.GetMainVM().IsToolGenerated = true;
-                //    doc.SetInput(point.X, point.Y);
-                //    //TODO Tool mode
-                //    // update collection
-                //    doc.GetMainVM().AddCollectionPoint(point);
-                //}
-
-                //if(doc.GetMainVM().ToolMode == CoordinateConversionLibrary.MapPointToolMode.Convert)
-                //    doc.GetMainVM().IsToolActive = false;
-            }
-            catch { }
-        }
-
-        protected override void OnMouseMove(MouseEventArgs arg)
-        {
-            try
-            {
-                IPoint point = GetMapPoint(arg.X, arg.Y);
-                ISnappingResult snapResult = null;
-                //Try to snap the current position
-                snapResult = m_Snapper.Snap(point);
-                m_SnappingFeedback.Update(snapResult, 0);
-                if (snapResult != null && snapResult.Location != null)
-                    point = snapResult.Location;
-
-                //TODO replace with Mediator call, MOUSE_MOVE_POINT
-                //var doc = AddIn.FromID<ArcMapAddinCoordinateConversion.DockableWindowCoordinateConversion.AddinImpl>(ThisAddIn.IDs.DockableWindowCoordinateConversion);
-
-                //if (doc != null && point != null)
-                //{
-                //    doc.GetMainVM().IsHistoryUpdate = false;
-                //    doc.SetInput(point.X, point.Y);
-                //}
-            }
-            catch { }
-        }
-
-        private IPoint GetMapPoint(int X, int Y)
-        {
-            //Get the active view from the ArcMap static class.
-            IActiveView activeView = ArcMap.Document.FocusMap as IActiveView;
-
-            var point = activeView.ScreenDisplay.DisplayTransformation.ToMapPoint(X, Y) as IPoint;
-
-            if (CoordinateConversionViewModel.AddInConfig.DisplayCoordinateType == CoordinateConversionLibrary.CoordinateTypes.None)
-            {
-                //IActiveView activeView = ArcMap.Document.FocusMap as IActiveView;
-                point.SpatialReference = ArcMap.Document.FocusMap.SpatialReference;
-            }
-            else
-            {
-                // always use WGS84
-                var sr = GetSR();
-
-                if (sr != null)
-                {
-                    point.Project(sr);
-                }
-            }
-
-            return point;
-        }
-
-        private ISpatialReference GetSR()
-        {
-            Type t = Type.GetTypeFromProgID("esriGeometry.SpatialReferenceEnvironment");
-            System.Object obj = Activator.CreateInstance(t);
-            ISpatialReferenceFactory srFact = obj as ISpatialReferenceFactory;
-
-            // Use the enumeration to create an instance of the predefined object.
-
-            IGeographicCoordinateSystem geographicCS =
-                srFact.CreateGeographicCoordinateSystem((int)
-                esriSRGeoCSType.esriSRGeoCS_WGS1984);
-
-            return geographicCS as ISpatialReference;
-        }
 
     }
